1.修复登录Api的data.money可能有小数导致转换异常的bug； 2.修复登录Api的data.next_exp有时返回数字有时返回字符导致的映射异常的bug； 3.其他细节小调整。

+++ src/Ray.BiliBiliTool.Agent/Dtos/CoinBalance.cs
@@ -5,6 +5,6 @@
     /// </summary>
     public class CoinBalance
     {
-        public int? Money { get; set; } = int.MinValue;
+        public decimal? Money { get; set; } = int.MinValue;
     }
 }
